moves profile creation automation to Welcome page and out of Settings page
