+++ src/pages/UserSettings/Settings.js
@@ -4,10 +4,8 @@ import { useNavigate } from "@reach/router";
 // Auth0 imports
 import { useAuth0 } from "../../config/react-auth0-spa";
 // Apollo/GraphQL imports
-import { useQuery, useMutation } from "react-apollo";
-
-// import ProfileForm from "./ProfileForm";
-import { ADD_USER_PROFILE, PROFILE_INFO } from "./queries";
+import { useQuery } from "react-apollo";
+import { PROFILE_INFO } from "./queries";
 
 // Material-UI imports
 import {
@@ -193,35 +191,13 @@ export default function Settings() {
   const classes = useStyles();
   const navigate = useNavigate();
   const userEmail = user.email;
-  const [createProfile] = useMutation(ADD_USER_PROFILE);
-
+  
   // Fetch profile for the user using the email associated with auth0 login
   const { loading, error, data, refetch } = useQuery(PROFILE_INFO, {
     variables: { email: user?.email },
   });
 
   const profile = data?.profile;
-
-  // Extract the profile from returning data of useQuery
-  useEffect(() => {
-    if (error) {
-      return <p>Error</p>;
-    }
-    // If user does not have a profile in backend, create one for them
-    if (!loading && !profile?.id) {
-      newProfile();
-    }
-    if (profile) {
-      refetch();
-    }
-    // eslint-disable-next-line
-  }, [profile]);
-
-  // Function that creates a profile for given email
-  const newProfile = async () => {
-    await createProfile({ variables: { email: user.email } });
-    refetch();
-  };
 
   return (
     <>
